Use gcc -E to output proper preprocessed OpenCL source
The thing is that (I don't know why, but the fact is) NVIDIA OpenCL
online compiler is not able to detect properly changes in includes
(probably compiler caches somewhere are stored?) and the old code
is used, even main render ray-tracing headers we changed heavily.
So use gcc to output preprocessed source. All problems seems
disappear.

The other part of this patch is to include OpenCL source as a
buffer, use xxd for generating header.